requirements.txt
Django==3.0.5
djangorestframework==3.11.0
psycopg2==2.7.3.2
six==1.14.0
requests==2.23.0
django-countries==6.1.2
praw==7.0.0
redis==3.5.2
celery==4.4.3
celery-progress==0.0.10
django-nose==1.4.6
selenium==3.141.0
behave==1.2.6
behave-django==1.4.0
beautifulsoup4==4.9.1
parse==1.15.0
parse-type==0.5.2
soupsieve==2.0.1
factory-boy==2.12.0
Faker==4.1.1
text-unidecode==1.3
coverage==5.0.3